requirements.txt
python-dotenv==1.1.0
requests==2.32.3
yt-dlp==2025.5.22
sentence-transformers==4.1.0
qdrant-client==1.14.2
nltk==3.9.1
textblob==0.19.0
python-docx==1.1.2
pdfplumber==0.11.6
openai==1.82.1
pymupdf==1.26.0
